Split client/server in to dedicated binaries
This primarily helps to clean up the command line.

Documentation updated, and ordered the sample config.json to match
documentation ordering.
